initial commit - slapd and TLS client
OpenLDAP provides slapd - an implementation of a hierarchical data store
that can be queried via LDAP protocol.

TLS auto-configuration is provided via tls-client relation and the
respective layer. A reference implementation of a provider for that is
easyrsa charm.

Configuration doesn't depend on config files which are deprecated - OLC
is used instead.